requirements.txt
pyyaml
requests
streamlit
web3
st_btn_select
pandas